+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Fixed namespace of some structures.
+
 ## [2.0.0] - 2024-06-07
 ### Changed
 - Update to support .Net 8 only.